AAP 51882 - shared jwt authenticator class (#874)
Currently, there are three component specific JWTConsumer classes in
DAB:

AWX:
https://github.com/ansible/django-ansible-base/blob/devel/ansible_base/jwt_consumer/awx/auth.py#L9
Hub:
https://github.com/ansible/django-ansible-base/blob/devel/ansible_base/jwt_consumer/hub/auth.py#L8
EDA:
https://github.com/ansible/django-ansible-base/blob/devel/ansible_base/jwt_consumer/eda/auth.py#L10

They all do the same, the goal of this PR is to provide one common
RbacEnabledJWTAuthentication class that all components use.

After this is merged, EDA, AWX and Hub need to be updated to use the
common class. Then we can remove the component specific classes in
another PR.